# Conflicts:
#	src/main/java/deco2800/skyfall/entities/AbstractEntity.java
#	src/main/java/deco2800/skyfall/saving/Save.java

+++ src/main/java/deco2800/skyfall/managers/database/InsertDataQueries.java
@@ -13,12 +13,12 @@ public class InsertDataQueries {
         this.statement = statement;
     }
 
-    public void insertSave(int id, String data) throws SQLException{
+    public void insertSave(long id, String data) throws SQLException{
         String query = String.format("insert into SAVES (save_id, data) values (%s, '%s')", id, data);
         statement.executeUpdate(query);
     }
 
-    public void insertWorld(int saveId, int worldId, boolean isCurrentWorld, String data) throws SQLException{
+    public void insertWorld(long saveId, long worldId, boolean isCurrentWorld, String data) throws SQLException{
         String query = String.format("insert into WORLDS (save_id, world_id, is_current_world, data) values (%s,%s,"
                 + "%s,'%s')", saveId, worldId, isCurrentWorld, data);
         statement.executeUpdate(query);
@@ -30,7 +30,7 @@ public class InsertDataQueries {
         statement.executeUpdate(query);
     }
 
-    public void insertNodes(int worldId, int xPos, int yPos, String data, int nodeId) throws SQLException{
+    public void insertNodes(long worldId, double xPos, double yPos, String data, long nodeId) throws SQLException{
         String query = String.format("insert into NODES (world_id, x_pos, y_pos, data, node_id) values (%s, %s, %s, "
                 + "'%s',%s)", worldId, xPos, yPos, data, nodeId);
         statement.executeUpdate(query);
